requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
catkin-pkg==1.0.0
certifi==2025.6.15
click==8.2.1
coloredlogs==15.0.1
distro==1.9.0
Django==5.2.3
docutils==0.21.2
fastapi==0.115.13
flatbuffers==25.2.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
image==1.5.33
jiter==0.10.0
mpmath==1.3.0
numpy==2.2.4
onnxruntime==1.21.0
openai==1.90.0
opencv-python==4.11.0.86
packaging==24.2
pillow==11.2.1
protobuf==6.30.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyrealsense2-macosx==2.54.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
rosdep==0.25.1
rosdistro==1.0.1
rosinstall==0.7.8
rosinstall-generator==0.1.23
rospkg==1.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
sympy==1.13.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
vcstools==0.1.42
wstool==0.1.17